+++ tests/integration/test_ourlogs.py
@@ -137,7 +137,7 @@ def test_ourlog_extraction_with_otel_logs(
         "organizationId": "1",
         "projectId": "42",
         "received": time_within_delta(),
-        "retentionDays": 90,
+        "retentionDays": 30,
         "serverSampleRate": 1.0,
         "timestamp": time_within_delta(
             ts, delta=timedelta(seconds=1), expect_resolution="ns"
@@ -312,7 +312,7 @@ def test_ourlog_extraction_with_sentry_logs(
             "organizationId": "1",
             "projectId": "42",
             "received": time_within_delta(),
-            "retentionDays": 90,
+            "retentionDays": 30,
             "serverSampleRate": 1.0,
             "timestamp": time_within_delta(
                 ts, delta=timedelta(seconds=1), expect_resolution="ns"
@@ -354,7 +354,7 @@ def test_ourlog_extraction_with_sentry_logs(
             "organizationId": "1",
             "projectId": "42",
             "received": time_within_delta(),
-            "retentionDays": 90,
+            "retentionDays": 30,
             "serverSampleRate": 1.0,
             "timestamp": time_within_delta(
                 ts, delta=timedelta(seconds=1), expect_resolution="ns"
@@ -628,7 +628,7 @@ def test_ourlog_extraction_default_pii_scrubbing_does_not_scrub_default_attribut
         "organizationId": "1",
         "projectId": "42",
         "received": time_within_delta(),
-        "retentionDays": 90,
+        "retentionDays": 30,
         "serverSampleRate": 1.0,
         "timestamp": time_within_delta(
             ts, delta=timedelta(seconds=1), expect_resolution="ns"
@@ -685,7 +685,7 @@ def test_ourlog_extraction_with_sentry_logs_with_missing_fields(
         "organizationId": "1",
         "projectId": "42",
         "received": time_within_delta(),
-        "retentionDays": 90,
+        "retentionDays": 30,
         "serverSampleRate": 1.0,
         "timestamp": time_within_delta(
             ts, delta=timedelta(seconds=1), expect_resolution="ns"
@@ -813,7 +813,7 @@ def test_browser_name_version_extraction(
         "organizationId": "1",
         "projectId": "42",
         "received": time_within_delta(),
-        "retentionDays": 90,
+        "retentionDays": 30,
         "serverSampleRate": 1.0,
         "timestamp": time_within_delta(
             ts, delta=timedelta(seconds=1), expect_resolution="ns"