+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+**Internal**:
+
+- Set the default log retention to 30 days for full fidelity and downsampled data. ([#5065](https://github.com/getsentry/relay/pull/5065))
 - Improved PII Scrubbing for attributes / logs ([#5061](https://github.com/getsentry/relay/pull/5061)))
 
 ## 25.8.0